Refactored Pages:
- Renamed poll_list.dart to boardroom.dart
- Deleted OLD boardrooms.dart
- Refactored home.dart, moving relevant methods to new boardroom.dart
